requirements.txt
alembic==1.6.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-API==2.0
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
flask-ngrok==0.0.25
Flask-Script==2.0.6
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.13
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
